+++ src/page/handover/note/NotePage.style.ts
@@ -18,18 +18,6 @@ export const noteSectionStyle = css({
   overflow: 'hidden',
 });
 
-export const entireInfoStyle = css({
-  display: 'flex',
-  flexDirection: 'column',
-
-  flexBasis: '34rem',
-  padding: '3.2rem 1.7rem 3.2rem 3.2rem',
-
-  borderRadius: '1.6rem 0 0 1.6rem',
-
-  backgroundColor: theme.colors.gray_100,
-});
-
 export const tabButtonStyle = css({
   width: '8.9rem',
 
@@ -41,38 +29,4 @@ export const infoContainerStyle = css({
   flexDirection: 'column',
 
   gap: '1.6rem',
-});
-
-export const infoLayoutStyle = css({
-  display: 'flex',
-
-  gap: '2rem',
-
-  alignItems: 'center',
-});
-
-export const titleStyle = css({
-  color: theme.colors.gray_800,
-
-  outline: 'none',
-  border: 'none',
-
-  whiteSpace: 'nowrap',
-
-  backgroundColor: theme.colors.gray_100,
-
-  marginBottom: '3.6rem',
-  ...theme.heading.heading03,
-
-  '::placeholder': {
-    color: theme.colors.gray_800,
-    ...theme.heading.heading03,
-  },
-});
-
-export const infoStyle = css({
-  width: '5.2rem',
-
-  color: theme.colors.gray_800,
-  whiteSpace: 'nowrap',
 });
