+++ src/page/handover/note/component/NoteDetail/NoteDetail.style.ts
@@ -1,0 +1,56 @@
+import { css } from '@emotion/react';
+
+import { theme } from '@/common/style/theme/theme';
+
+export const entireInfoStyle = css({
+  display: 'flex',
+  flexDirection: 'column',
+
+  flexBasis: '34rem',
+  padding: '3.2rem 1.7rem 3.2rem 3.2rem',
+
+  borderRadius: '1.6rem 0 0 1.6rem',
+
+  backgroundColor: theme.colors.gray_100,
+});
+
+export const titleStyle = css({
+  color: theme.colors.gray_800,
+
+  outline: 'none',
+  border: 'none',
+
+  whiteSpace: 'nowrap',
+
+  backgroundColor: theme.colors.gray_100,
+
+  marginBottom: '3.6rem',
+  ...theme.heading.heading03,
+
+  '::placeholder': {
+    color: theme.colors.gray_800,
+    ...theme.heading.heading03,
+  },
+});
+
+export const infoContainerStyle = css({
+  display: 'flex',
+  flexDirection: 'column',
+
+  gap: '1.6rem',
+});
+
+export const infoLayoutStyle = css({
+  display: 'flex',
+
+  gap: '2rem',
+
+  alignItems: 'center',
+});
+
+export const infoStyle = css({
+  width: '5.2rem',
+
+  color: theme.colors.gray_800,
+  whiteSpace: 'nowrap',
+});
